PR Feedback changes including option to not read encap in recorder, spacing fixes, and new file to hold various template helper functions
