+++ dds/DCPS/XTypes/TypeObject.h
@@ -828,7 +828,7 @@ namespace XTypes {
   // ID of a type member
   typedef ACE_CDR::ULong MemberId;
   const ACE_CDR::ULong MEMBER_ID_INVALID = UINT32_MAX;
-  // Implementation specific sentinel for a union discriminator used in DynamicData
+  /// Implementation specific sentinel for a union discriminator used in DynamicData
   const ACE_CDR::ULong DISCRIMINATOR_ID = UINT32_MAX - 1;
   const ACE_CDR::ULong ANNOTATION_STR_VALUE_MAX_LEN = 128;
   const ACE_CDR::ULong ANNOTATION_OCTETSEC_VALUE_MAX_LEN = 128;
