+++ tests/DCPS/XTypes/DynamicTypes/Pub/XTypesDynamicPub.cpp
@@ -13,8 +13,8 @@
 #include <dds/DCPS/transport/framework/TransportRegistry.h>
 #include <dds/DCPS/transport/framework/TransportSendStrategy.h>
 #if defined ACE_AS_STATIC_LIBS && !defined OPENDDS_SAFETY_PROFILE
-#include <dds/DCPS/RTPS/RtpsDiscovery.h>
-#include <dds/DCPS/transport/rtps_udp/RtpsUdp.h>
+#  include <dds/DCPS/RTPS/RtpsDiscovery.h>
+#  include <dds/DCPS/transport/rtps_udp/RtpsUdp.h>
 #endif
 
 #include <ace/Arg_Shifter.h>
@@ -93,9 +93,8 @@ void union_default_narrow_write(DataWriter_var dw)
   narrow_dw->write(foo, handle);
 }
 
-int ACE_TMAIN(int argc, ACE_TCHAR *argv[])
+int ACE_TMAIN(int argc, ACE_TCHAR* argv[])
 {
-  int status = 0;
   const ACE_TCHAR* type_name = argv[1];
   DomainParticipantFactory_var dpf = TheParticipantFactoryWithArgs(argc, argv);
   DomainParticipant_var dp =
@@ -166,5 +165,5 @@ int ACE_TMAIN(int argc, ACE_TCHAR *argv[])
   dpf->delete_participant(dp);
   TheServiceParticipant->shutdown();
 
-  return status;
+  return 0;
 }
