+++ tests/DCPS/XTypes/DynamicTypes/Pub/dynamic.idl
@@ -43,7 +43,7 @@ struct outer_struct {
 };
 
 @topic
-@extensibility(MUTABLE)
+@mutable
 union my_union switch (long) {
   case 1: long double ld;
   case 2: bool_seq my_alias_seq;
